refactor(look&feel,apollo): simplify BasePicture component
