+++ client/apollo/css/src/BasePicture/BasePicture.scss
@@ -1,0 +1,5 @@
+.af-basepicture {
+  width: calc(48 / var(--font-size-base) * 1rem);
+  height: calc(48 / var(--font-size-base) * 1rem);
+  border-radius: 100%;
+}
