--- client/apollo/css/src/BasePicture/BasePictureApollo.scss
@@ -1,1 +1,0 @@
-@use "./BasePictureCommon";
