--- client/apollo/css/src/BasePicture/BasePictureCommon.scss
@@ -1,5 +1,0 @@
-.af-basepicture {
-  width: 48px;
-  height: 48px;
-  border-radius: 100vmax;
-}
