--- client/apollo/css/src/BasePicture/BasePictureLF.scss
@@ -1,1 +1,0 @@
-@use "./BasePictureCommon";
