+++ client/apollo/css/src/apollo.scss
@@ -14,7 +14,7 @@
 @use "./ClickIcon/ClickIconApollo";
 @use "./Link/LinkApollo";
 @use "./Toggle/ToggleApollo";
-@use "./BasePicture/BasePictureApollo";
+@use "./BasePicture/BasePicture";
 @use "./Message/MessageApollo";
 @use "./Tag/TagApollo";
 @use "./CardMessage/CardMessageApollo";
