+++ client/apollo/css/src/look-and-feel.scss
@@ -14,7 +14,7 @@
 @use "./ClickIcon/ClickIconLF";
 @use "./Link/LinkLF";
 @use "./Toggle/ToggleLF";
-@use "./BasePicture/BasePictureLF";
+@use "./BasePicture/BasePicture";
 @use "./Message/MessageLF";
 @use "./Tag/TagLF";
 @use "./Accordion/AccordionLF";
