--- client/apollo/react/src/BasePicture/BasePictureApollo.tsx
@@ -1,3 +1,0 @@
-import "@axa-fr/design-system-apollo-css/dist/BasePicture/BasePictureApollo.scss";
-
-export { BasePicture } from "./BasePictureCommon";
