--- client/apollo/react/src/BasePicture/BasePictureLF.tsx
@@ -1,3 +1,0 @@
-import "@axa-fr/design-system-apollo-css/dist/BasePicture/BasePictureLF.scss";
-
-export { BasePicture } from "./BasePictureCommon";
