+++ client/apollo/react/src/BasePicture/__tests__/BasePicture.test.tsx
@@ -1,7 +1,7 @@
 import logo from "@axa-fr/design-system-apollo-css/logo-axa.svg";
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import { BasePicture } from "../BasePictureCommon";
+import { BasePicture } from "../BasePicture";
 
 describe("BasePicture component", () => {
   it("should render with default class name", () => {
