+++ client/apollo/react/src/ContentItemMono/ContentItemMonoApollo.tsx
@@ -1,15 +1,10 @@
 import "@axa-fr/design-system-apollo-css/dist/ContentItemMono/ContentItemMonoApollo.scss";
-import { BasePicture } from "../BasePicture/BasePictureApollo";
 import { Icon } from "../Icon/IconApollo";
 import {
   ContentItemMonoCommon,
-  ContentItemProps,
+  type ContentItemProps,
 } from "./ContentItemMonoCommon";
 
 export const ContentItemMono = (props: ContentItemProps) => (
-  <ContentItemMonoCommon
-    {...props}
-    BasePictureComponent={BasePicture}
-    IconComponent={Icon}
-  />
+  <ContentItemMonoCommon {...props} IconComponent={Icon} />
 );
