+++ client/apollo/react/src/ContentItemMono/ContentItemMonoLF.tsx
@@ -1,15 +1,10 @@
 import "@axa-fr/design-system-apollo-css/dist/ContentItemMono/ContentItemMonoLF.scss";
-import { BasePicture } from "../BasePicture/BasePictureLF";
 import { Icon } from "../Icon/IconLF";
 import {
   ContentItemMonoCommon,
-  ContentItemProps,
+  type ContentItemProps,
 } from "./ContentItemMonoCommon";
 
 export const ContentItemMono = (props: ContentItemProps) => (
-  <ContentItemMonoCommon
-    {...props}
-    BasePictureComponent={BasePicture}
-    IconComponent={Icon}
-  />
+  <ContentItemMonoCommon {...props} IconComponent={Icon} />
 );
