+++ client/apollo/react/src/List/ClickItem/components/ClickItemPrefixApollo.tsx
@@ -1,14 +1,9 @@
 import { Icon } from "../../../Icon/IconApollo";
-import { BasePicture } from "../../../BasePicture/BasePictureApollo";
 import {
   ClickItemPrefixCommon,
   type ClickItemPrefixProps,
 } from "./ClickItemPrefixCommon";
 
 export const ClickItemPrefix = (props: ClickItemPrefixProps) => (
-  <ClickItemPrefixCommon
-    BasePictureComponent={BasePicture}
-    {...props}
-    IconComponent={Icon}
-  />
+  <ClickItemPrefixCommon {...props} IconComponent={Icon} />
 );
