+++ client/apollo/react/src/List/ClickItem/components/ClickItemPrefixCommon.tsx
@@ -1,7 +1,7 @@
-import { ComponentProps, ComponentType } from "react";
-import { BasePicture } from "../../../BasePicture/BasePictureCommon";
+import type { ComponentProps, ComponentType } from "react";
+import { BasePicture } from "../../../BasePicture/BasePicture";
 import type { IconProps } from "../../../Icon/IconCommon";
-import { ClickItemStates, ClickItemVariants } from "../ClickItemCommon";
+import type { ClickItemStates, ClickItemVariants } from "../ClickItemCommon";
 
 export type ClickItemPrefixProps = {
   state: ClickItemStates;
@@ -12,7 +12,6 @@ export type ClickItemPrefixProps = {
 
 export type ClickItemPrefixCommonProps = ClickItemPrefixProps & {
   IconComponent: ComponentType<IconProps>;
-  BasePictureComponent: ComponentType<ComponentProps<typeof BasePicture>>;
 };
 
 export const ClickItemPrefixCommon = ({
@@ -21,14 +20,13 @@ export const ClickItemPrefixCommon = ({
   icon,
   basePictureProps,
   IconComponent,
-  BasePictureComponent,
 }: ClickItemPrefixCommonProps) => {
   if (!(Boolean(basePictureProps) && variant === "agent") && !icon) {
     return null;
   }
 
   if (Boolean(basePictureProps) && variant === "agent") {
-    return <BasePictureComponent {...basePictureProps} />;
+    return <BasePicture {...basePictureProps} />;
   }
 
   return (
