+++ client/apollo/react/src/List/ClickItem/components/ClickItemPrefixLF.tsx
@@ -1,14 +1,9 @@
 import { Icon } from "../../../Icon/IconLF";
-import { BasePicture } from "../../../BasePicture/BasePictureLF";
 import {
   ClickItemPrefixCommon,
   type ClickItemPrefixProps,
 } from "./ClickItemPrefixCommon";
 
 export const ClickItemPrefix = (props: ClickItemPrefixProps) => (
-  <ClickItemPrefixCommon
-    BasePictureComponent={BasePicture}
-    {...props}
-    IconComponent={Icon}
-  />
+  <ClickItemPrefixCommon {...props} IconComponent={Icon} />
 );
